Add performance comparison script and threshold constants
- Introduced `profile_comparison.sh` script for benchmarking Rust and C implementations of RIES.
- Added `thresholds.rs` module to centralize named constants for numerical comparisons, improving code readability and maintainability.
- Created property-based tests in `property_tests.rs` to verify mathematical properties of expression evaluation and derivatives using proptest.
- Included regression seeds in `property_tests.proptest-regressions` for consistent testing of previously failing cases.

+++ src/metrics.rs
@@ -19,7 +19,7 @@ pub struct MatchMetrics {
     /// Whether this is an exact match (error < 1e-14)
     pub is_exact: bool,
     /// Total complexity score
-    pub complexity: u16,
+    pub complexity: u32,
     /// "Ugliness" penalty (deep nesting, many ops)
     pub ugliness: f64,
     /// Novelty score (rarer operators/constants)
